requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
fakeredis>=2.20.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
